rename views to controllers
